+++ odoo/local-src/specific_zetes/tests/test_catchweight_parking.py
@@ -12,13 +12,12 @@ class TestCatchweightParking(ZetesParkingTest):
 
         report_query = """
         SELECT report.id
-        FROM report_stock_quant_bylocation AS report
+        FROM report_stock_refill_arrange AS report
           LEFT JOIN stock_location ON stock_location.id = report.location_id
           LEFT JOIN picking_zone
             ON stock_location.picking_zone_id = picking_zone.id
         WHERE report.product_id = %s
           AND picking_zone.code = %s
-        ORDER BY report.refill_priority
         LIMIT 1
         """
         self.env.cr.execute(report_query, (self.product_1.id,
@@ -28,10 +27,10 @@ class TestCatchweightParking(ZetesParkingTest):
         self.assertTrue(result)
         report_id = result[0]
 
-        model_name = 'report.stock.quant.bylocation'
+        model_name = 'report.stock.refill.arrange'
         report = self.env[model_name].browse(report_id)
         # Create the picking
-        self.picking_parking = report.create_parking_picking()
+        self.picking_parking = report.create_picking()
 
         self.reserve_medicament = self.env['stock.location'].create({
             'name': 'GD80F4',