+++ HISTORY.rst
@@ -8,6 +8,7 @@ Unreleased
 
 **Features and Improvements**
 
+* Refactor refill report (arrange, reassort) for real-time computation
 * ALCYN-1761: improve the view of the action to send invoices by email / print, make it less confusing
 * ALCYN-1761: generate and send pdf invoices in background
 * ALCYN-1761: the "Related" button on the job related to invoices now try to open the correct view (customer or supplier)